requirements.txt
requests==2.32.3
line-bot-sdk==3.14.2
flask==3.1.0
geopy==2.4.1